refactor: infra의 status와 domain의 status 분리
