+++ src/main/java/roomescape/payment/application/OrderScheduler.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Component;
 import roomescape.payment.application.dto.OrderInfo;
 import roomescape.payment.application.dto.PaymentResult;
 import roomescape.payment.domain.Order;
-import roomescape.payment.domain.OrderStatus;
+import roomescape.payment.domain.PaymentStatus;
 import roomescape.payment.infra.client.exception.TossBusinessException.PaymentNotFound;
 import roomescape.reservation.application.ReservationManager;
 import roomescape.reservation.application.ReservationReader;
@@ -93,7 +93,7 @@ public class OrderScheduler {
     }
 
     private void checkTossPayDone(Order order, PaymentResult tossStatus) {
-        if (tossStatus.status() == OrderStatus.COMPLETED) {
+        if (tossStatus.status() == PaymentStatus.COMPLETED) {
             log.warn("[CRITICAL] 미결제로 방치된 주문이 토스에서는 결제 성공 상태입니다! 즉시 자동 환불합니다. OrderId: {}", order.getOrderId());
             paymentService.cancelBySystem(order.getOrderId(), "결제 완료 후 시스템 지연/이탈로 인한 자동 환불");
         }
@@ -103,7 +103,7 @@ public class OrderScheduler {
         try {
             OrderInfo order = orderService.getOrder(pending.getId());
 
-            if (order.status() == OrderStatus.COMPLETED) {
+            if (order.status() == PaymentStatus.COMPLETED) {
                 paymentService.cancelBySystem(order.orderId(), "예약일 경과로 인한 대기 예약 자동 환불");
                 log.info("환불 성공: 예약ID={}, 주문ID={}", pending.getId(), order.orderId());
             }
