+++ src/main/java/roomescape/payment/application/dto/PaymentResult.java
@@ -2,13 +2,13 @@ package roomescape.payment.application.dto;
 
 import java.time.LocalDateTime;
 import lombok.Builder;
-import roomescape.payment.domain.OrderStatus;
+import roomescape.payment.domain.PaymentStatus;
 
 @Builder
 public record PaymentResult(
         String paymentKey,
         String orderId,
-        OrderStatus status,
+        PaymentStatus status,
         Long approvedAmount,
         LocalDateTime createdAt
 ) {
