--- src/main/java/roomescape/payment/infra/client/PaymentStatus.java
@@ -1,24 +1,0 @@
-package roomescape.payment.infra.client;
-
-public enum PaymentStatus {
-    READY,
-    IN_PROGRESS,
-    WAITING_FOR_DEPOSIT,
-    DONE,
-    CANCELED,
-    PARTIAL_CANCELED,
-    ABORTED,
-    EXPIRED,
-    UNKNOWN;
-
-    public static PaymentStatus from(String raw) {
-        if (raw == null) {
-            return UNKNOWN;
-        }
-        try {
-            return valueOf(raw);
-        } catch (IllegalArgumentException e) {
-            return UNKNOWN;
-        }
-    }
-}
